// magic-ai-app/types.ts
export type Membership = 'free' | 'trial' | 'amateur' | 'semi-pro' | 'professional' | 'expired';

export interface User {
  id?: string;
  email: string;
  membership: Membership;
  isAdmin?: boolean;
  trialEndDate?: number; 
  generationCount: number;
  lastResetDate: string; // ISO String
  emailVerified?: boolean;
}
// ...
export interface IllusionPrinciple {
  name: string;
  description: string;
}

export interface IllusionBlueprintResponse {
  potential_principles: IllusionPrinciple[];
  blueprint_description: string;
}